Mini script to correctly bind to cpus from within an mpirun call
The script allows to pass information on the cpu set where the
ranks should be executing. This allows to achieve something along
the line of

  $ mpirun -np N taskset -c <cpu_list> <exec_name> <exec_args>

The advantage of the script is that the script itself takes care
of specifying a different cpu_list for each rank, so that each
rank remains pinned to a core.